feature: Odin inspector, refactor, model

+++ Assets/_Project/Scripts/Events/PlayerSignals.cs
@@ -6,18 +6,18 @@
 
     public readonly struct SignalChangeSpeedAndAnimation: ISignalChangeSpeed,ISignalChangeAnimation
     { 
-        public SignalChangeSpeedAndAnimation(string animation, int speed)
+        public SignalChangeSpeedAndAnimation(string animation, float speed)
         {
             Animation = animation;
             Speed = speed;
         }
     
         public string Animation { get; }
-        public int Speed { get; }
+        public float Speed { get; }
     } 
     public interface ISignalChangeSpeed
     {
-        int  Speed { get; }
+        float  Speed { get; }
     } 
     public interface ISignalChangeAnimation
     {
